pyproject.toml
[tool.poetry]
name = "vector-intro"
version = "0.1.0"
description = ""
authors = ["Darlington02 <darlingtonnnam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.0"
langchain = "^0.3.7"
langchain-openai = "^0.2.6"
python-dotenv = "^1.0.1"
langchainhub = "^0.1.21"
black = "^24.10.0"
langchain-pinecone = "^0.2.0"
langchain-community = "^0.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
